Final Deliverables/requirements.txt
numpy==1.23.3
opencv_python==4.6.0.66
playsound==1.3.0
python-dotenv==0.21.0
tensorflow==2.10.0
twilio==7.15.0
